Moved hash tagging analysis plugin out of interface and added tests
